Salt theme has several layers and there are base variables that can simplify the overrides.
- added theme contract for `interact` color palette
- SaltMenuItem has a hardcoded hex value, so we needed a global style to correct

This change ensures the VerticalNavigation uses the Mosaic `selectable` colors and not the default
Pro Sidebar colors
